[TASK] replace the listDir functionality with `find` command
[DESC.]
- add `maxDepth` option to list API
- refactoring `fs-manager.js`
     - separated into `fs-service.js`(access file system) and `Resource.js`(model)
     - remove needless functions
     - clean up some duplicated methods
